refactor: amend root ./src/index.ts exports in @observerly/fits
refactor: amend root ./src/index.ts exports in @observerly/fits
