+++ src/index.ts
@@ -6,22 +6,7 @@
 
 /*****************************************************************************************************************/
 
-// Data Unit
-export * from './dataunit'
-
-/*****************************************************************************************************************/
-
 // Flexible Image Transport System (FITS) Header
-export * from './header'
-
-/*****************************************************************************************************************/
-
-// Flexible Image Transport System (FITS)
-export * from './fits'
-
-/*****************************************************************************************************************/
-
-// Utilities
-export * from './utilities'
+export type { FITSHeader } from './header'
 
 /*****************************************************************************************************************/
